fix(common): fix type errors in test helper (#320)
* clean up test helper WIP

* improve types

* update snapshot
