+++ packages/cli/src/commands/__tests__/add.test.ts
@@ -48,6 +48,7 @@ describe('add', () => {
         "index.html",
         "package.json",
         "src",
+        "test",
         "tsconfig.json",
         "tsconfig.node.json",
         "vite.config.ts",
@@ -87,6 +88,7 @@ describe('add', () => {
         "index.html",
         "package.json",
         "src",
+        "test",
         "tsconfig.json",
         "tsconfig.node.json",
         "vite.config.ts",
